+++ NEWS.md
@@ -15,7 +15,7 @@
   * A possible `.n` column now contains `0` instead of `NA` in case all values of a certain temporal aggregation level are missing.
   * The stripping of missing values within summary functions now depends on the value of a possible `na.rm` argument.
 * Fixed the support for `data.table`'s *GForce* optimisation when the `fun` argument of the `aggregate()` method is provided with a character vector specifying summary functions. Please note that the column order of the resulting `DTSg` object is now different due to this fix.
-* Fixed that the `set()` method was too anxious about removing all value columns.
+* Fixed that the `setCols()` method and its R6 alias `set()` was too anxious about removing all value columns.
 * It is no longer possible to use the deprecated `byFasttime*()` TALFs. Please use the `"fasttime"` or even better the new `"timechange"` `funbyApproach` from now on.
 * Removed `magrittr` from the suggested packages list (vignettes now use R's native pipe operator).
 * Slightly improved the documentation.