+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * More time zones are now considered UTC or equivalent (execute `grep("^(Etc/)?(UTC|UCT|Universal|Zulu)$|^(Etc/)?(GMT(\\+|-)?0?|Greenwich)$", OlsonNames(), ignore.case = TRUE, value = TRUE)` for a full list).
 * Slightly improved the documentation.
+* Minor internal code improvements, which require R ≥ 4.0.0.
 
 # DTSg v2.0.0